refactor common retagprompt
Introduce a new source file common.py for common commands and, as a
first example, define RetagPromptCommand there and register it from
search.py.

This shows how to "decorate" a class that is defined somewhere else, and
thus how to decorate it multiple times.
